refactor(/repository): 분리되었던 테스트 한 클래스에 통합
